Core/Guilds: Fix few uint32 -> uint64 fields for money

+++ src/server/game/Scripting/ScriptMgr.cpp
@@ -1109,7 +1109,7 @@ void ScriptMgr::OnGuildEvent(Guild* p_Guild, uint8 p_EventType, uint32 p_PlayerG
 /// @p_ItemOrMoney    : Item entry or gold amount
 /// @p_ItemStackCount : Item stack count
 /// @p_DestTabID      : Destination tab ID
-void ScriptMgr::OnGuildBankEvent(Guild* p_Guild, uint8 p_EventType, uint8 p_TabID, uint32 p_PlayerGUID, uint32 p_ItemOrMoney, uint16 p_ItemStackCount, uint8 p_DestTabID)
+void ScriptMgr::OnGuildBankEvent(Guild* p_Guild, uint8 p_EventType, uint8 p_TabID, uint32 p_PlayerGUID, uint64 p_ItemOrMoney, uint16 p_ItemStackCount, uint8 p_DestTabID)
 {
     FOREACH_SCRIPT(GuildScript)->OnBankEvent(p_Guild, p_EventType, p_TabID, p_PlayerGUID, p_ItemOrMoney, p_ItemStackCount, p_DestTabID);
 }
